Revert "Several improvements."
