+++ app/ConvexClientProvider.tsx
@@ -12,9 +12,7 @@ export default function ConvexClientProvider({
   children: ReactNode;
 }) {
   return (
-    <ClerkProvider
-      publishableKey={process.env.NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY!}
-    >
+    <ClerkProvider publishableKey="pk_test_bGVnYWwtY291Z2FyLTMxLmNsZXJrLmFjY291bnRzLmRldiQ">
       <ConvexProviderWithClerk client={convex} useAuth={useAuth}>
         {children}
       </ConvexProviderWithClerk>
