--- app/auth.ts
@@ -1,5 +1,0 @@
-import { auth } from '@clerk/nextjs';
-
-export async function getAuthToken() {
-  return (await auth().getToken({ template: 'convex' })) ?? undefined;
-}
