+++ app/recording/[id]/page.tsx
@@ -1,21 +1,25 @@
+'use client';
+
 import Header from '@/components/ui/Header';
 import RecordingDesktop from '@/components/pages/recording/RecordingDesktop';
 import RecordingMobile from '@/components/pages/recording/RecordingMobile';
 import { api } from '@/convex/_generated/api';
-import { preloadQuery } from 'convex/nextjs';
-import { getAuthToken } from '@/app/auth';
-import RecordingPage from './recording';
+import { useQuery } from 'convex/react';
 
-const Page = async ({ params }: { params: { id: string } }) => {
+const QuickCheckPage = ({ params }: { params: { id: string } }) => {
   const id = params.id as any;
-  const token = await getAuthToken();
-  const preloadedNote = await preloadQuery(
-    api.notes.getNote,
-    { id },
-    { token },
-  );
+  const currentNote = useQuery(api.notes.getNote, { id });
+  if (!currentNote) return null; // Some 404 page maybe
 
-  return <RecordingPage preloadedNote={preloadedNote} />;
+  return (
+    <div className="">
+      <Header />
+      <div className="mx-auto max-w-[1500px]">
+        <RecordingDesktop {...currentNote} />
+        <RecordingMobile {...currentNote} />
+      </div>
+    </div>
+  );
 };
 
-export default Page;
+export default QuickCheckPage;
