+++ components/ui/Header.tsx
@@ -3,10 +3,8 @@
 import { useUser } from '@clerk/clerk-react';
 import Link from 'next/link';
 import { UserNav } from './UserNav';
-import { useConvexAuth } from 'convex/react';
 
 const Header = () => {
-  const { isLoading } = useConvexAuth();
   const { user } = useUser();
 
   return (
@@ -27,9 +25,7 @@ const Header = () => {
         </Link>
         {/* buttons */}
         <div className="flex w-fit items-center gap-[22px]">
-          {isLoading ? (
-            <></>
-          ) : user ? (
+          {user ? (
             <>
               <Link
                 href={'/dashboard'}
