+++ components/ui/UserNav.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { LogOut, PlusCircle, Paintbrush2 } from 'lucide-react';
 import { Avatar, AvatarFallback, AvatarImage } from './avatar';
 import { Button } from './button';
