+++ convex/_generated/api.d.ts
@@ -4,7 +4,7 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.9.0.
+ * Generated by convex@1.6.2.
  * To regenerate, run `npx convex dev`.
  * @module
  */
@@ -13,11 +13,10 @@ import type {
   ApiFromModules,
   FilterApi,
   FunctionReference,
-} from 'convex/server';
-import type * as notes from '../notes.js';
-import type * as openai from '../openai.js';
-import type * as utils from '../utils.js';
-import type * as whisper from '../whisper.js';
+} from "convex/server";
+import type * as notes from "../notes.js";
+import type * as openai from "../openai.js";
+import type * as whisper from "../whisper.js";
 
 /**
  * A utility for referencing Convex functions in your app's API.
@@ -30,14 +29,13 @@ import type * as whisper from '../whisper.js';
 declare const fullApi: ApiFromModules<{
   notes: typeof notes;
   openai: typeof openai;
-  utils: typeof utils;
   whisper: typeof whisper;
 }>;
 export declare const api: FilterApi<
   typeof fullApi,
-  FunctionReference<any, 'public'>
+  FunctionReference<any, "public">
 >;
 export declare const internal: FilterApi<
   typeof fullApi,
-  FunctionReference<any, 'internal'>
+  FunctionReference<any, "internal">
 >;
