+++ convex/_generated/api.js
@@ -4,12 +4,12 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.9.0.
+ * Generated by convex@1.6.2.
  * To regenerate, run `npx convex dev`.
  * @module
  */
 
-import { anyApi } from 'convex/server';
+import { anyApi } from "convex/server";
 
 /**
  * A utility for referencing Convex functions in your app's API.
