+++ convex/_generated/dataModel.d.ts
@@ -4,7 +4,7 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.9.0.
+ * Generated by convex@1.6.2.
  * To regenerate, run `npx convex dev`.
  * @module
  */
@@ -14,9 +14,9 @@ import type {
   DocumentByName,
   TableNamesInDataModel,
   SystemTableNames,
-} from 'convex/server';
-import type { GenericId } from 'convex/values';
-import schema from '../schema.js';
+} from "convex/server";
+import type { GenericId } from "convex/values";
+import schema from "../schema.js";
 
 /**
  * The names of all of your Convex tables.
