+++ convex/_generated/server.js
@@ -4,7 +4,7 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.9.0.
+ * Generated by convex@1.6.2.
  * To regenerate, run `npx convex dev`.
  * @module
  */
@@ -17,7 +17,7 @@ import {
   internalActionGeneric,
   internalMutationGeneric,
   internalQueryGeneric,
-} from 'convex/server';
+} from "convex/server";
 
 /**
  * Define a query in this Convex app's public API.
