+++ package.json
@@ -17,8 +17,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
-    "convex": "^1.9",
-    "convex-helpers": "^0.1.15",
+    "convex": "^1.6.2",
     "lucide-react": "^0.292.0",
     "next": "^14.0.4",
     "openai": "^4.20.1",
